Combine 'post-confirm' with 'confirm' step
