+++ client/lib/signup/step-actions/index.js
@@ -926,18 +926,21 @@ export function excludeStepIfEmailVerified( stepName, defaultDependencies, nextP
 		return;
 	}
 
-	if ( ! nextProps.isEmailVerified ) {
-		return;
-	}
-
-	let wasSkipped = true;
+	/* For the P2 signup flow, if we displayed the email verification step before,
+	   we need to display it again when the user comes back to the flow
+	   after verification. */
 	if (
 		nextProps.flowName === 'p2-new' &&
 		nextProps?.progress[ stepName ]?.status === 'in-progress'
 	) {
-		wasSkipped = false;
+		return;
 	}
-	nextProps.submitSignupStep( { stepName, wasSkipped } );
+
+	if ( ! nextProps.isEmailVerified ) {
+		return;
+	}
+
+	nextProps.submitSignupStep( { stepName, wasSkipped: true } );
 	flows.excludeStep( stepName );
 }
