+++ client/signup/config/step-components.js
@@ -69,7 +69,6 @@ const stepNameToModuleName = {
 	'p2-site': 'p2-site',
 	'p2-get-started': 'p2-get-started',
 	'p2-confirm-email': 'p2-confirm-email',
-	'p2-email-confirmed': 'p2-email-confirmed',
 	'plans-business-monthly': 'plans',
 	'plans-ecommerce-monthly': 'plans',
 	'plans-personal-monthly': 'plans',
