+++ client/signup/config/steps.js
@@ -20,7 +20,6 @@ import {
 	isNewOrExistingSiteFulfilled,
 	setDesignIfNewSite,
 	excludeStepIfEmailVerified,
-	excludeStepIfEmailVerificationSkipped,
 } from 'calypso/lib/signup/step-actions';
 import { generateSteps } from './steps-pure';
 
@@ -46,5 +45,4 @@ export default generateSteps( {
 	isNewOrExistingSiteFulfilled,
 	setDesignIfNewSite,
 	excludeStepIfEmailVerified,
-	excludeStepIfEmailVerificationSkipped,
 } );
